+++ backend/serializers/database_check_serializer.py
@@ -1,5 +1,5 @@
 from rest_framework import serializers
-from models.database_models import DatabaseCheckModel
+from custom_models.database_models import DatabaseCheckModel
 
 # Serializers define the API representation
 class DatabaseCheckSerializer(serializers.ModelSerializer):
